+++ src/renderer/src/App.jsx
@@ -9,28 +9,28 @@ import Dashboard from './Pages/Dashboard/Dashboard'
 import RootLayout from './layout/RootLayout/RootLayout'
 import Settings from './Pages/Settings/Settings'
 // Converters
-import JsonYml from './Pages/Converters/json_yml'
-import NumberBase from './Pages/Converters/number_base'
-import Timestamp from './Pages/Converters/timestamp'
+import JsonYml from './Pages/Converters/JsonYml'
+import NumberBase from './Pages/Converters/NumberBase'
+import Timestamp from './Pages/Converters/Timestamp'
 // Encoders and Decoders
-import Base64 from './Pages/Encoders_Decoders/base64_text'
-import EdHtml from './Pages/Encoders_Decoders/html'
-import EdJwt from './Pages/Encoders_Decoders/jwt'
-import EdUrl from './Pages/Encoders_Decoders/url'
+import Base64 from './Pages/Encoders_Decoders/Base64Text'
+import EdHtml from './Pages/Encoders_Decoders/Html'
+import EdJwt from './Pages/Encoders_Decoders/Jwt'
+import EdUrl from './Pages/Encoders_Decoders/Url'
 // Generators
-import Checksum from './Pages/Generators/checksum'
-import Hash from './Pages/Generators/hash'
-import RandomText from './Pages/Generators/random_text'
-import SecretKeys from './Pages/Generators/secret_keys'
-import Uuid from './Pages/Generators/uuid'
+import Checksum from './Pages/Generators/Checksum'
+import Hash from './Pages/Generators/Hash'
+import RandomText from './Pages/Generators/RandomText'
+import SecretKeys from './Pages/Generators/SecretKeys'
+import Uuid from './Pages/Generators/Uuid'
 // Graphic
-import ColorPicker from './Pages/Graphic/color_picker'
+import ColorPicker from './Pages/Graphic/ColorPicker'
 // Text
-import CaseManipulator from './Pages/Text/case_manipulator'
-import MarkdownPreview from './Pages/Text/markdown_preview'
-import Regex from './Pages/Text/regex'
+import CaseManipulator from './Pages/Text/CaseManipulator'
+import MarkdownPreview from './Pages/Text/MarkdownPreview'
+import Regex from './Pages/Text/Regex'
 // Api Testing
-import Http from './Pages/Api_Testing/http'
+import Http from './Pages/Api_Testing/Http'
 import './App.css'
 
 function App() {
