+++ src/renderer/src/Pages/Generators/RandomText.jsx
