+++ src/renderer/src/Pages/Text/CaseManipulator.jsx
