+++ src/renderer/src/utils/APITestingUtils/utils/helpers.js
@@ -1,14 +1,14 @@
 const convertKeyValueToObject = (keyPairs) => {
   return [...keyPairs].reduce((data, pair) => {
-    const key = pair.keyItem;
-    const value = pair.valueItem;
+    const key = pair.keyItem
+    const value = pair.valueItem
 
-    if (key === '') return data;
+    if (key === '') return data
     return {
       ...data,
-      [key]: value,
-    };
-  }, {});
-};
+      [key]: value
+    }
+  }, {})
+}
 
-export { convertKeyValueToObject };
+export { convertKeyValueToObject }
