+++ src/renderer/src/utils/APITestingUtils/utils/useStateCallback.js
@@ -1,20 +1,20 @@
-import { useState, useCallback, useRef, useEffect } from 'react';
+import { useState, useCallback, useRef, useEffect } from 'react'
 
 export default function useStateCallback(initialState) {
-  const [state, setState] = useState(initialState);
-  const cbRef = useRef(null);
+  const [state, setState] = useState(initialState)
+  const cbRef = useRef(null)
 
   const setStateCallback = useCallback((state, cb) => {
-    cbRef.current = cb;
-    setState(state);
-  }, []);
+    cbRef.current = cb
+    setState(state)
+  }, [])
 
   useEffect(() => {
     if (cbRef.current) {
-      cbRef.current(state);
-      cbRef.current = null;
+      cbRef.current(state)
+      cbRef.current = null
     }
-  }, [state]);
+  }, [state])
 
-  return [state, setStateCallback];
+  return [state, setStateCallback]
 }
